pyproject.toml
[tool.poetry]
name = "did_sdk_py"
version = "0.0.1"
description = "The repository contains the Python SDK for managing DID Documents and Anoncreds Verifiable Credentials registry using Hedera Consensus Service."
authors = ["Alexander Shenshin <alexander.shenshin@dsr-corporation.com>"]
repository = "https://github.com/hashgraph/did-sdk-py"
documentation = "https://hashgraph.github.io/did-sdk-py/"
readme = "README.md"
packages = [
    { include = "did_sdk_py" }
]

[tool.poetry.dependencies]
aiohttp = "^3.10.9"
aiohttp_retry = "^2.9.0"
python = ">=3.10,<4.0"
pyjnius = ">=1.5.0"
did-resolver = "0.0.3"
base58 = "^2.1.1"
hedera-sdk-py = { git = "https://github.com/DSRCorporation/hedera-sdk-py", branch = "feature/update-java-sdk" }
zstandard = "^0.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-anything = "^0.1.4"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
pytest-retry = "^1.6.3"
deptry = "^0.16.2"
pyright = "1.1.386"
ruff = "^0.7.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = { extras = ["python"], version = "^0.26.1" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



[tool.pyright]
include = ["did_sdk_py"]
venvPath = "."
venv = ".venv"
executionEnvironments = [
    { root = "did_sdk_py" }
]

typeCheckingMode = "standard"
reportInvalidTypeForm = "warning"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    #"TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["did_sdk_py"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
